DL/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
cycler==0.10.0
importlib-metadata==4.6.1
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.0
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-mnist==0.7
six==1.16.0
toml==0.10.2
torch==1.9.0
typing-extensions==3.10.0.0
zipp==3.5.0
